[refactor] Moved electricity into the widgets package
