+++ lib/widgets/electricty.dart
